Query renders but code needs cleaning up
