+++ dashboard/src/actions/index.js
@@ -3,7 +3,8 @@ import {
     checkStatus,
     parseJSON,
     isNotEmpty,
-    showTreeView, // isShortestPath,
+    showTreeView,
+    processGraph, // isShortestPath,
 } from "../containers/Helpers";
 
 const addQuery = text => ({
@@ -21,6 +22,47 @@ const saveErrorResponse = text => ({
     type: "ERROR_RESPONSE",
     text,
 });
+
+const saveResponseProperties = obj => ({
+    type: "RESPONSE_PROPERTIES",
+    ...obj,
+});
+
+const renderGraph = (query, result, treeView, dispatch) => {
+    let startTime = new Date();
+
+    let [nodes, edges, labels, nodesIdx, edgesIdx] = processGraph(
+        result,
+        treeView,
+        query,
+    );
+
+    let endTime = new Date(),
+        timeTaken = (endTime.getTime() - startTime.getTime()) / 1000,
+        render = "";
+
+    if (timeTaken > 1) {
+        render = timeTaken.toFixed(1) + "s";
+    } else {
+        render = (timeTaken - Math.floor(timeTaken)) * 1000 + "ms";
+    }
+
+    dispatch(
+        saveResponseProperties({
+            plotAxis: labels,
+            allNodes: nodes,
+            allEdges: edges,
+            numNodes: nodes.length,
+            numEdges: edges.length,
+            nodes: nodes.slice(0, nodesIdx),
+            edges: edges.slice(0, edgesIdx),
+            treeView: treeView,
+            latency: result.server_latency.total,
+            data: result,
+            rendering: render,
+        }),
+    );
+};
 
 export const runQuery = query => {
     return dispatch => {
@@ -43,8 +85,9 @@ export const runQuery = query => {
                     dispatch(saveSuccessResponse(JSON.stringify(result)));
                 } else if (isNotEmpty(result)) {
                     dispatch(addQuery(query));
-                    // let mantainSortOrder = showTreeView(query);
+                    let mantainSortOrder = showTreeView(query);
                     dispatch(saveSuccessResponse(null, result));
+                    renderGraph(query, result, mantainSortOrder, dispatch);
                 } else {
                     dispatch(
                         saveErrorResponse(
