+++ dashboard/src/containers/App.js
@@ -138,213 +138,6 @@ function hasProperties(props: Object): boolean {
   return Object.keys(props).length !== 1;
 }
 
-function processGraph(response: Object, treeView: boolean, query: string) {
-  let nodesStack: Array<ResponseNode> = [],
-    // Contains map of a lable to its shortform thats displayed.
-    predLabel: MapOfStrings = {},
-    // Map of whether a Node with an Uid has already been created. This helps
-    // us avoid creating duplicating nodes while parsing the JSON structure
-    // which is a tree.
-    uidMap: MapOfBooleans = {},
-    nodes: Array<Node> = [],
-    edges: Array<Edge> = [],
-    emptyNode: ResponseNode = {
-      node: {},
-      src: {
-        id: "",
-        pred: "empty",
-      },
-    },
-    someNodeHasChildren: boolean = false,
-    ignoredChildren: Array<ResponseNode> = [],
-    // We store the indexes corresponding to what we show at first render here.
-    // That we can only do one traversal.
-    nodesIndex,
-    edgesIndex,
-    level = 0;
-
-  for (var root in response) {
-    if (!response.hasOwnProperty(root)) {
-      continue;
-    }
-
-    someNodeHasChildren = false;
-    ignoredChildren = [];
-    if (root !== "server_latency" && root !== "uids") {
-      let block = response[root];
-      for (let i = 0; i < block.length; i++) {
-        let rn: ResponseNode = {
-          node: block[i],
-          src: {
-            id: "",
-            pred: root,
-          },
-        };
-        if (hasChildren(block[i])) {
-          someNodeHasChildren = true;
-          nodesStack.push(rn);
-        } else {
-          ignoredChildren.push(rn);
-        }
-      }
-
-      // If no node has children or its a shortest path query, then we add root
-      // level nodes to the view.
-      if (!someNodeHasChildren) {
-        nodesStack.push.apply(nodesStack, ignoredChildren);
-      }
-    }
-  }
-
-  // We push an empty node after all the children. This would help us know when
-  // we have traversed all nodes at a level.
-  nodesStack.push(emptyNode);
-
-  while (nodesStack.length > 0) {
-    let obj = nodesStack.shift();
-    // Check if this is an empty node.
-    if (Object.keys(obj.node).length === 0 && obj.src.pred === "empty") {
-      if (level === 0 || level === 1 && nodes.length < 100) {
-        nodesIndex = nodes.length;
-        edgesIndex = edges.length;
-      }
-
-      if (nodesStack.length === 0) {
-        break;
-      }
-
-      nodesStack.push(emptyNode);
-      level++;
-      continue;
-    }
-
-    let properties: MapOfStrings = {
-      attrs: {},
-      facets: {},
-    },
-      hasChildNodes: boolean = false,
-      id: string,
-      edgeAttributes = {
-        facets: {},
-      },
-      uid: string;
-
-    uid = obj.node["_uid_"] === undefined ? uuid() : obj.node["_uid_"];
-    id = treeView
-      ? // For tree view, the id is the join of ids of this node
-        // with all its ancestors. That would make it unique.
-        [obj.src.id, uid].filter(val => val).join("-")
-      : uid;
-
-    for (let prop in obj.node) {
-      if (!obj.node.hasOwnProperty(prop)) {
-        continue;
-      }
-
-      // We can have a key-val pair, another array or an object here (in case of facets)
-      let val = obj.node[prop];
-      if (Array.isArray(val)) {
-        hasChildNodes = true;
-        let arr = val, xposition = 1;
-        for (let j = 0; j < arr.length; j++) {
-          arr[j]["x"] = xposition++;
-          nodesStack.push({
-            node: arr[j],
-            src: {
-              pred: prop,
-              id: id,
-            },
-          });
-        }
-      } else if (typeof val === "object") {
-        if (prop === "@facets") {
-          // lets handle @facets between uids here.
-          for (let pred in val) {
-            if (!val.hasOwnProperty(pred)) {
-              continue;
-            }
-
-            // pred could either be _ or other predicates. If its a predicate it could have
-            // multiple k-v pairs.
-            if (pred === "_") {
-              edgeAttributes["facets"] = val["_"];
-            } else {
-              let predFacets = val[pred];
-              for (let f in predFacets) {
-                if (!predFacets.hasOwnProperty(f)) {
-                  continue;
-                }
-
-                properties["facets"][`${pred}[${f}]`] = predFacets[f];
-              }
-            }
-          }
-        }
-      } else {
-        properties["attrs"][prop] = val;
-      }
-    }
-
-    if (!hasProperties(obj) && !hasChildNodes) {
-      continue;
-    }
-
-    let nodeAttrs = properties["attrs"];
-    let aggrTerm = aggregationPrefix(nodeAttrs);
-    let name = aggrTerm !== "" ? aggrTerm : obj.src.pred;
-
-    let props = getGroupProperties(name, predLabel, groups);
-    let x = nodeAttrs["x"];
-    delete nodeAttrs["x"];
-
-    let n: Node = {
-      id: id,
-      x: x,
-      label: getNodeLabel(nodeAttrs),
-      title: JSON.stringify(properties),
-      color: props.color,
-      group: obj.src.pred,
-    };
-
-    if (!uidMap[id]) {
-      // For tree view, we can't push duplicates because two query blocks might have the
-      // same root node and child elements won't really have the same uids as their uid is a
-      // combination of all their ancestor uids.
-      uidMap[id] = true;
-      nodes.push(n);
-    }
-
-    if (obj.src.id === "") {
-      continue;
-    }
-
-    var e: Edge = {
-      from: obj.src.id,
-      to: id,
-      title: JSON.stringify(edgeAttributes),
-      label: props.label,
-      color: props.color,
-      arrows: "to",
-    };
-    edges.push(e);
-  }
-
-  var plotAxis = [];
-  for (let pred in groups) {
-    if (!groups.hasOwnProperty(pred)) {
-      continue;
-    }
-
-    plotAxis.push({
-      label: groups[pred]["label"],
-      pred: pred,
-      color: groups[pred]["color"],
-    });
-  }
-
-  return [nodes, edges, plotAxis, nodesIndex, edgesIndex];
-}
-
 type QueryTs = {|
   text: string,
   lastRun: number,
@@ -532,48 +325,12 @@ class App extends React.Component {
     groups = {};
   };
 
-  renderGraph = (result, treeView) => {
-    let startTime = new Date(), that = this;
-
-    let [nodes, edges, labels, nodesIdx, edgesIdx] = processGraph(
-      result,
-      treeView,
-      this.state.lastQuery,
-    );
-
-    that.setState({
-      plotAxis: labels,
-      allNodes: nodes,
-      allEdges: edges,
-      nodes: nodes.slice(0, nodesIdx),
-      edges: edges.slice(0, edgesIdx),
-      treeView: treeView,
-    });
-
-    let endTime = new Date(),
-      timeTaken = (endTime.getTime() - startTime.getTime()) / 1000,
-      render = "";
-
-    if (timeTaken > 1) {
-      render = timeTaken.toFixed(1) + "s";
-    } else {
-      render = (timeTaken - Math.floor(timeTaken)) * 1000 + "ms";
-    }
-
-    that.setState({
-      latency: result.server_latency.total,
-      resType: "",
-      result: result,
-      rendering: render,
-    });
-  };
-
-  renderResText = (type, text) => {
-    this.setState({
-      resType: type,
-      response: text,
-    });
-  };
+  // renderResText = (type, text) => {
+  //   this.setState({
+  //     resType: type,
+  //     response: text,
+  //   });
+  // };
 
   render = () => {
     return (
